requirements.txt
Flask==0.10.1
Flask-Admin==1.0.7
Flask-HTTPAuth==2.2.0
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-Principal==0.4.0
Flask-Security==1.6.9
Flask-WTF==0.9.3
Jinja2==2.7
MarkupSafe==0.18
Pillow==2.6.1
PyYAML==3.11
SQLAlchemy==0.8.4
Theano==0.6.0
WTForms==1.0.5
Werkzeug==0.8.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
blinker==1.3
cairocffi==0.6
certifi==14.05.14
cffi==0.9.2
ez-setup==0.9
flask-mongoengine==0.7.0
gizeh==0.1.10
gnureadline==6.3.3
greenlet==0.4.5
ipdb==0.8
ipython==2.3.1
itsdangerous==0.23
lxml==3.2.4
matplotlib==1.4.2
mock==1.0.1
mongoengine==0.8.6
mongokit==0.9.0
nose==1.3.4
numpy==1.9.2
passlib==1.6.1
psycopg2==2.5.1
py-bcrypt==0.4
pycparser==2.12
pymongo==2.6.3
pyparsing==2.0.3
pysbp==0.0.2
python-dateutil==2.4.0
pytz==2014.10
six==1.9.0
tornado==4.1
wsgiref==0.1.2
